move category related types to utils
