+++ frontend/src/components/StudentProfile/PointsForm/types.ts
@@ -1,10 +1,4 @@
-import { CategoriesQuery } from "../../../graphql/categories.graphql.types";
-
 export type FormPoints = {
   subcategoryId: string;
   points: number;
 };
-
-export type Subcategory = NonNullable<
-  CategoriesQuery["categories"][number]["subcategories"][number]
->;
