+++ frontend/src/utils/utils.ts
@@ -4,6 +4,18 @@ type User = {
   role: string;
 };
 
+export type Category = {
+  id: string;
+  name: string;
+  subcategories: Subcategory[];
+};
+
+export type Subcategory = {
+  id: string;
+  name: string;
+  maxPoints: number;
+};
+
 export function hasRole(user: User, allowedRoles: Roles[]) {
   return allowedRoles.includes(user.role as Roles);
 }
